Implemented Dijkstra's algorithm for displaying path finding.
After researching I believe Dijkstra is better for highlighting all reachable tiles, while Astar is better for focusing the best path to a specific tile and is faster to load. But because I don't have a specific tile I need to go to, I will use Dijkstra, however I'll probably update it and/or change it after.